Live inventory (#144)
* Rip out old inventory system.

* Readd inventory timing & charging neatly.

* Update upcoming to include the newly announce Alpha 5.2.

* Dirty fix for bukkit's lack of a proper InventoryMoveItem with player inventories.

* Deprecate API methods that use account double values.
- These will most likely be kept until 6.0 then removed in 6.1

* Switch to BigDecimal. Add config for decimal places limit on currency formatting.
- Major step towards sponge support
- Allows us to control decimal places easier

* Currency decimal places kept are now configurable.

* Start work on issue #129

* Plan out database redesign

* A few fixes.

* Push modifications for archival purposes.

* New inventory structure, still have to optimize the move item hack for it.

* Fixes for Minecraft 1.7 servers.

* Paid players now receive a message.

* Banks now support multiple currencies.
- This is a much needed feature for being feature-complete towards multiple currencies
- This has also changed how interest is configured, it's not per-currency.

* Forge Mod Mob Reward Support & ability to rename mobs via messages.yml

* Fixes for latest additions.

* Ability to specify a currency for mob rewards.

* Updates files and changelog for 5.2

* worlds.yml support for mob currency rewards

* Money Top Command, with baltop and balancetop shortcuts

* World Sharing revamp, Fix for configurations not loading for shared worlds.

* Added slime size support for mob rewards.

* We take precedence over Essentials in terms of economy handling.

* Mob Rewards Additions
- Ability to allow for a chance reward, which gives a random reward between a set of two amounts
- Ability to give a custom message for various mobs
- Ability to give a custom reward for mobs with a custom name
- Ability to have a Multiplier per mob, player, world, or just a general multiplier which will multiple a mob's reward by a given amount.

* Update Upcoming.md for new mob rewards additions.

* Add getter for API, and instance variables.

* Add items.yml for sign shops.

* Added ability to disable economy features for specific worlds

* Reduced MaterialHelper size by over 400 lines.

* Changed some configuration defaults

* Lots of Sign Shop progress
What's left to implement?
 - The transactions part, which should only require two more lines.

* Sign Shops are now complete.

* Some polish on the previous commit.

* Ability to use ender chest inventory for shop signs, and general fixes.

* Final SignShop Polish

* Inventory Research:
Hopefully this will save other developers the headache I've dealt with working on checking when an item has been moved.

* Tidy Inventory Research Doc

* Inventory Research for shift clicks.

* Inventory Work.

* ChestInventory & Bank Signs

* Fixes for factions & item currencies #142.

* Make ItemShops compatible with 1.7

* More decimal places for item currencies

* More Inventory Final Touches.
- There's weird issues with shift clicks, and collecting to cursor that still need fixed.

* Polish Inventory Actions.
- Alpha 5.2 may have limited action click

+++ src/com/github/tnerevival/core/objects/MaterialObject.java
@@ -1,18 +1,20 @@
 package com.github.tnerevival.core.objects;
 
+import java.math.BigDecimal;
+
 public class
 MaterialObject {
 
   private String name;
   private Boolean item;
-  private Double cost;
-  private Double value;
-  private Double use;
-  private Double crafting;
-  private Double enchant;
-  private Double place;
-  private Double mine;
-  private Double smelt;
+  private BigDecimal cost;
+  private BigDecimal value;
+  private BigDecimal use;
+  private BigDecimal crafting;
+  private BigDecimal enchant;
+  private BigDecimal place;
+  private BigDecimal mine;
+  private BigDecimal smelt;
 
   public MaterialObject(String name) {
     this.name = name;
@@ -34,67 +36,67 @@ MaterialObject {
     this.item = item;
   }
 
-  public Double getCost() {
+  public BigDecimal getCost() {
     return cost;
   }
 
-  public void setCost(Double cost) {
+  public void setCost(BigDecimal cost) {
     this.cost = cost;
   }
 
-  public Double getValue() {
+  public BigDecimal getValue() {
     return value;
   }
 
-  public void setValue(Double value) {
+  public void setValue(BigDecimal value) {
     this.value = value;
   }
 
-  public Double getUse() {
+  public BigDecimal getUse() {
     return use;
   }
 
-  public void setUse(Double use) {
+  public void setUse(BigDecimal use) {
     this.use = use;
   }
 
-  public Double getCrafting() {
+  public BigDecimal getCrafting() {
     return crafting;
   }
 
-  public void setCrafting(Double crafting) {
+  public void setCrafting(BigDecimal crafting) {
     this.crafting = crafting;
   }
 
-  public Double getEnchant() {
+  public BigDecimal getEnchant() {
     return enchant;
   }
 
-  public void setEnchant(Double enchant) {
+  public void setEnchant(BigDecimal enchant) {
     this.enchant = enchant;
   }
 
-  public Double getPlace() {
+  public BigDecimal getPlace() {
     return place;
   }
 
-  public void setPlace(Double place) {
+  public void setPlace(BigDecimal place) {
     this.place = place;
   }
 
-  public Double getMine() {
+  public BigDecimal getMine() {
     return mine;
   }
 
-  public void setMine(Double mine) {
+  public void setMine(BigDecimal mine) {
     this.mine = mine;
   }
 
-  public Double getSmelt() {
+  public BigDecimal getSmelt() {
     return smelt;
   }
 
-  public void setSmelt(Double smelt) {
+  public void setSmelt(BigDecimal smelt) {
     this.smelt = smelt;
   }
 }
